Fix transcription failure: switch encoder dtype from fp32 to fp16
Root cause found (user hit this on a company laptop with the exact
error): the fp32 encoder for whisper-large-v3-turbo (and kotoba-whisper)
exceeds 2GB, which forces onnxruntime-web to load it via its "external
data" (.onnx_data) path. That path is broken in the bundled
onnxruntime-web version and fails with:

  Deserialize tensor onnx::MatMul_3964 failed. Failed to load external
  data file "encoder_model.onnx_data", error: Module.MountedFiles is
  not available.

This is a known transformers.js issue (huggingface/transformers.js#989),
with the fix confirmed by the maintainer: use a dtype under 2GB so
external data is never needed. Switched encoder_model dtype to fp16
(verified both onnx-community/whisper-large-v3-turbo and
onnx-community/kotoba-whisper-v2.2-ONNX publish an encoder_model_fp16.onnx).
This was very likely the actual cause of the "transcription never
finishes" reports, not a WebGPU shader-compile stall.

Bumped to 0.4.1 (bug fix, no new capability).

+++ web/src/transcribe-worker.ts
@@ -55,7 +55,11 @@ async function getPipeline(
   cachedPipeline = await createPipeline("automatic-speech-recognition", modelId, {
     device,
     dtype: {
-      encoder_model: "fp32",
+      // fp32 encoder exceeds 2GB, forcing onnxruntime-web's external-data (.onnx_data) loading
+      // path, which is broken for these models (fails with "Module.MountedFiles is not
+      // available" — see https://github.com/huggingface/transformers.js/issues/989). fp16
+      // keeps the encoder under 2GB and avoids external data entirely.
+      encoder_model: "fp16",
       decoder_model_merged: "q4",
     },
     progress_callback: (data: { status: string; file?: string; progress?: number }) => {
